refactor(fe): extract DangerZoneSection
